refactor:  update transaction detail modal and utility functions for improved transaction handling
